% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_module.R
\name{analyse_module}
\alias{analyse_module}
\title{Visualization of gene expression modules in samples with Heatmap}
\usage{
analyse_module(
  diffGenes,
  modulePick,
  moduleAssign,
  heatcolors,
  toExport,
  filename
)
}
\arguments{
\item{diffGenes}{voom-processed DGEList}

\item{modulePick}{Integer of specific module to analyse from \code{\link[stats]{cutree}} function}

\item{moduleAssign}{Output of \code{\link[stats]{cutree}} function}

\item{heatcolors}{RColorBrewer palette}

\item{toExport}{Boolean indicating whether a .tsv file of the module genes should be exported}

\item{filename}{filename of exported .tsv file, if toExport is set to TRUE}
}
\description{
Wrapper function to analyse gene modules
}
