% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLGSP_metadata}
\alias{BLGSP_metadata}
\title{Sample metadata for each patient sample}
\format{
A tibble containing 129 observations of 20 variables:
\describe{
\item{cohort}{Discovery/validation/Tonsils}
\item{patient_barcode}{Patient barcode named according to OCG project code CGCI nomenclature structure}
\item{clinical variant}{endemic/sporadic BL}
...
}
}
\source{
\url{https://cgci-data.nci.nih.gov/Public/BLGSP/clinical/SFU/metadata.patients.tsv}
}
\usage{
BLGSP_metadata
}
\description{
Sample metadata for each patient sample
}
\keyword{datasets}
