% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broadSet.C2.ALL}
\alias{broadSet.C2.ALL}
\title{MSigDB C2 collection}
\format{
A GeneSetCollection object from MSigDB:
\describe{
\item{geneID()}{Extract gene IDs in each signature}
...
}
}
\source{
\url{"https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp"}
}
\usage{
broadSet.C2.ALL
}
\description{
MSigDB C2 collection
}
\keyword{datasets}
