% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{Split string and select substring by index}
\usage{
clean_name(name, split_by, element, start, stop)
}
\arguments{
\item{name}{A character string}

\item{split_by}{Character to split string by}

\item{element}{Element index to select from substring}

\item{start}{Start index of substring to select}

\item{stop}{End index of substring to select}
}
\value{
An indexed substring
}
\description{
Helper function to generate the shortened sample names (replacing the longer biospecimen_id names)
}
