% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_mds.R
\name{col_mds}
\alias{col_mds}
\title{Generate colors for \code{\link[limma]{plotMDS}} function from Rcolorbrewer palettes}
\usage{
col_mds(metadata, column_name, brewer_pal = "Set1")
}
\arguments{
\item{metadata}{The metadata dataframe with column containing group of interest}

\item{column_name}{A string of the column name}

\item{brewer_pal}{The Rcolorbrewer palette to select colors from}
}
\value{
A character vector containing color names
}
\description{
Helper function used for plotting the multidimensional scaling (MDS) plot
}
