% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebv_genes}
\alias{ebv_genes}
\title{List of EBV genes}
\format{
A list containing 113 EBV genes:
\describe{
\item{LMP2}{LMP2 gene}
...
}
}
\source{
\url{"https://ftp.ncbi.nih.gov/genomes/Viruses/human_gammaherpesvirus_4_uid14413/NC_007605.gff"}
}
\usage{
ebv_genes
}
\description{
List of EBV genes
}
\keyword{datasets}
