% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gostplot.R
\name{plot_gostplot}
\alias{plot_gostplot}
\title{Plot the Manhattan plot for functional enrichment results from \code{\link[gprofiler2]{gostplot}}}
\usage{
plot_gostplot(diffGenes, modulePick, moduleAssign, interactive = F)
}
\arguments{
\item{diffGenes}{voom-processed DGEList}

\item{modulePick}{Integer of specific module to analyse from \code{\link[stats]{cutree}} function}

\item{moduleAssign}{Output of \code{\link[stats]{cutree}} function}

\item{interactive}{Boolean. If TRUE, will return an interactive gostplot.}
}
\description{
Plot the Manhattan plot for functional enrichment results from \code{\link[gprofiler2]{gostplot}}
}
