% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log2cpm.R
\name{plot_log2cpm}
\alias{plot_log2cpm}
\title{Plot the log2 counts per million of samples}
\usage{
plot_log2cpm(DGEList, samples_id, subtitle)
}
\arguments{
\item{DGEList}{An edgeR::DGEList object}

\item{samples_id}{A grouped_df data.frame with a biospecimen_id column}

\item{subtitle}{Subtitle name for ggplot}
}
\description{
Plot the log2 counts per million of samples
}
